Switched from int to floating point

+++ src/main/java/rocks/zipcode/atm/CashMachine.java
@@ -39,7 +39,7 @@ public class CashMachine {
         return accountActive;
     }
 
-    public void deposit(int amount) {
+    public void deposit(double amount) {
         if (accountData != null) {
             tryCall(
                     () -> bank.deposit(accountData, amount),
@@ -48,7 +48,7 @@ public class CashMachine {
         }
     }
 
-    public void withdraw(int amount) {
+    public void withdraw(double amount) {
         if (accountData != null) {
             tryCall(
                     () -> bank.withdraw(accountData, amount),
